second

+++ src/main/java/com/driver/service/impl/OrderServiceImpl.java
@@ -19,8 +19,7 @@ public class OrderServiceImpl implements OrderService {
 
     @Autowired
     OrderRepository orderRepository;
-    @Autowired
-    UserRepository userRepository;
+
 
     @Override
     public OrderDto createOrder(OrderDto orderDto) {
@@ -32,14 +31,7 @@ public class OrderServiceImpl implements OrderService {
         orderEntity.setItems(orderDto.getItems());
         orderEntity.setStatus(true);
 
-        //now by saving the user in repository the order will also be saved as we have done bidirectional mapping
-        UserEntity userEntity = userRepository.findByUserId(orderDto.getUserId());
-        userEntity.getOrderEntityList().add(orderEntity);
-        userRepository.save(userEntity);
-
-        // extracting the user again to get updated user
-        userEntity = userRepository.findByUserId(orderDto.getUserId());
-        orderEntity = userEntity.getOrderEntityList().get(userEntity.getOrderEntityList().size() - 1);
+        orderRepository.save(orderEntity);
 
         // setting the orderDto to return it
         orderDto.setOrderId(orderEntity.getOrderId());
@@ -51,11 +43,17 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public OrderDto getOrderById(String orderId) throws Exception {
+
+        if(orderId.equals("")){
+            throw new Exception();
+        }
+
         OrderEntity orderEntity = orderRepository.findByOrderId(orderId);
 
         if(orderEntity == null){
             throw new Exception();
         }
+
         // putting the values of order into orderDto and returning it
         OrderDto orderDto = new OrderDto();
         orderDto.setId(orderEntity.getId());
@@ -70,6 +68,11 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public OrderDto updateOrderDetails(String orderId, OrderDto orderDto) throws Exception {
+
+        if(orderId.equals("") || orderDto == null){
+            throw new Exception();
+        }
+
         OrderEntity orderEntity = orderRepository.findByOrderId(orderId);
 
         if(orderEntity == null){
@@ -83,7 +86,6 @@ public class OrderServiceImpl implements OrderService {
         orderRepository.save(orderEntity);
 
         // extracting the order details from the repository to get updated details
-        orderEntity = orderRepository.findByOrderId(orderId);
         orderDto.setOrderId(orderEntity.getOrderId());
         orderDto.setId(orderEntity.getId());
         orderDto.setUserId(orderEntity.getUserId());
@@ -93,20 +95,23 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public void deleteOrder(String orderId) throws Exception {
+        if(orderId.equals("")){
+            throw new Exception();
+        }
+
         OrderEntity orderEntity = orderRepository.findByOrderId(orderId);
+
         if(orderEntity == null){
             throw new Exception();
         }
-        UserEntity userEntity = userRepository.findByUserId(orderEntity.getUserId());
-        userEntity.getOrderEntityList().remove(orderEntity);
-        userRepository.save(userEntity);
+
         orderRepository.delete(orderEntity);
     }
 
     @Override
     public List<OrderDto> getOrders() {
 
-        List<OrderEntity> orderEntityList = orderRepository.findAllOrders();
+        List<OrderEntity> orderEntityList = (List<OrderEntity>) orderRepository.findAll();
         List<OrderDto> orderDtoList = new ArrayList<>();
 
         for(OrderEntity orderEntity : orderEntityList){
